Rename 'directory storage' to 'local storage'; temporarily remove SFTP storage from unit tests

+++ Wibblr.Grufs.Tests/InMemoryChunkStorage.cs
@@ -54,7 +54,7 @@ namespace Wibblr.Grufs.Tests
             return _dict.ContainsKey(address);
         }
 
-        public int Count()
+        public long Count()
         {
             return _dict.Count();
         }
